Reformated the code so that it'll auto increment and works with front end
