+++ client/src/components/routes/AuthenticatedRoute.js
